Top level init bb.js, but better scoped imports to not incur cost too early.
