+++ yarn-project/circuits.js/src/barretenberg/crypto/aes128/index.test.ts
@@ -1,5 +1,3 @@
-import { init } from '@aztec/foundation/crypto';
-
 import { createCipheriv, createDecipheriv, randomBytes } from 'crypto';
 
 import { Aes128 } from './index.js';
@@ -7,8 +5,7 @@ import { Aes128 } from './index.js';
 describe('aes128', () => {
   let aes128!: Aes128;
 
-  beforeAll(async () => {
-    await init();
+  beforeAll(() => {
     aes128 = new Aes128();
   });
