+++ yarn-project/circuits.js/src/barretenberg/crypto/grumpkin/index.test.ts
@@ -1,4 +1,3 @@
-import { init } from '@aztec/foundation/crypto';
 import { createDebugLogger } from '@aztec/foundation/log';
 
 import { GrumpkinScalar, Point } from '../../../index.js';
@@ -9,8 +8,7 @@ const debug = createDebugLogger('bb:grumpkin_test');
 describe('grumpkin', () => {
   let grumpkin!: Grumpkin;
 
-  beforeAll(async () => {
-    await init();
+  beforeAll(() => {
     grumpkin = new Grumpkin();
   });
