+++ yarn-project/cli/src/test/utils.test.ts
@@ -5,7 +5,8 @@ import { InvalidArgumentError } from 'commander';
 import { MockProxy, mock } from 'jest-mock-extended';
 
 import { encodeArgs } from '../encoding.js';
-import { getTxSender, parseSaltFromHexString, stripLeadingHex } from '../utils.js';
+import { parseSaltFromHexString } from '../parse_args.js';
+import { getTxSender, stripLeadingHex } from '../utils.js';
 import { mockContractArtifact } from './mocks.js';
 
 describe('CLI Utils', () => {
