+++ yarn-project/foundation/src/abi/utils.ts
@@ -1,4 +1,4 @@
-import { ABIType } from './abi.js';
+import { type ABIType } from './abi.js';
 
 /**
  * Returns whether the ABI type is an Aztec or Ethereum Address defined in Aztec.nr.
